Pale Moon 19.0.1

+++ build/mobile/remoteautomation.py
@@ -3,19 +3,23 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import time
-import sys
+import re
 import os
-import socket
 import automationutils
 import tempfile
 import shutil
+import subprocess
 
 from automation import Automation
-from devicemanager import DeviceManager, NetworkTools
+from devicemanager import NetworkTools, DMError
+
+# signatures for logcat messages that we don't care about much
+fennecLogcatFilters = [ "The character encoding of the HTML document was not declared",
+                           "Use of Mutation Events is deprecated. Use MutationObserver instead." ]
 
 class RemoteAutomation(Automation):
     _devicemanager = None
-    
+
     def __init__(self, deviceManager, appName = '', remoteLog = None):
         self._devicemanager = deviceManager
         self._appName = appName
@@ -24,11 +28,12 @@ class RemoteAutomation(Automation):
 
         # Default our product to fennec
         self._product = "fennec"
+        self.lastTestSeen = "remoteautomation.py"
         Automation.__init__(self)
 
     def setDeviceManager(self, deviceManager):
         self._devicemanager = deviceManager
-        
+
     def setAppName(self, appName):
         self._appName = appName
 
@@ -37,7 +42,7 @@ class RemoteAutomation(Automation):
 
     def setProduct(self, product):
         self._product = product
-        
+
     def setRemoteLog(self, logfile):
         self._remoteLog = logfile
 
@@ -62,37 +67,54 @@ class RemoteAutomation(Automation):
 
         return env
 
-    def waitForFinish(self, proc, utilityPath, timeout, maxTime, startTime, debuggerInfo, symbolsDir):
+    def waitForFinish(self, proc, utilityPath, timeout, maxTime, startTime, debuggerInfo, symbolsPath):
+        """ Wait for tests to finish (as evidenced by the process exiting),
+            or for maxTime elapse, in which case kill the process regardless.
+        """
         # maxTime is used to override the default timeout, we should honor that
         status = proc.wait(timeout = maxTime)
-
-        print proc.stdout
+        self.lastTestSeen = proc.getLastTestSeen
 
         if (status == 1 and self._devicemanager.processExist(proc.procName)):
             # Then we timed out, make sure Fennec is dead
+            if maxTime:
+                print "TEST-UNEXPECTED-FAIL | %s | application ran for longer than " \
+                      "allowed maximum time of %s seconds" % (self.lastTestSeen, maxTime)
+            else:
+                print "TEST-UNEXPECTED-FAIL | %s | application ran for longer than " \
+                      "allowed maximum time" % (self.lastTestSeen)
             proc.kill()
 
         return status
 
     def checkForCrashes(self, directory, symbolsPath):
-        dumpDir = tempfile.mkdtemp()
-        self._devicemanager.getDirectory(self._remoteProfile + '/minidumps/', dumpDir)
-        automationutils.checkForCrashes(dumpDir, symbolsPath, self.lastTestSeen)
-        try:
-          shutil.rmtree(dumpDir)
-        except:
-          print "WARNING: unable to remove directory: %s" % (dumpDir)
+        remoteCrashDir = self._remoteProfile + '/minidumps/'
+        if self._devicemanager.dirExists(remoteCrashDir):
+            dumpDir = tempfile.mkdtemp()
+            self._devicemanager.getDirectory(remoteCrashDir, dumpDir)
+            automationutils.checkForCrashes(dumpDir, symbolsPath,
+                                            self.lastTestSeen)
+            try:
+                shutil.rmtree(dumpDir)
+            except:
+                print "WARNING: unable to remove directory: %s" % dumpDir
+        else:
+            # As of this writing, the minidumps directory is automatically
+            # created when fennec (first) starts, so its lack of presence
+            # is a hint that something went wrong.
+            print "WARNING: No crash directory (%s) on remote " \
+                "device" % remoteCrashDir
 
     def buildCommandLine(self, app, debuggerInfo, profileDir, testURL, extraArgs):
         # If remote profile is specified, use that instead
         if (self._remoteProfile):
             profileDir = self._remoteProfile
 
-        # Hack for robocop, if app & testURL == None and extraArgs contains the rest of the stuff, lets 
+        # Hack for robocop, if app & testURL == None and extraArgs contains the rest of the stuff, lets
         # assume extraArgs is all we need
         if app == "am" and extraArgs[0] == "instrument":
             return app, extraArgs
- 
+
         cmd, args = Automation.buildCommandLine(self, app, debuggerInfo, profileDir, testURL, extraArgs)
         # Remove -foreground if it exists, if it doesn't this just returns
         try:
@@ -113,13 +135,14 @@ class RemoteAutomation(Automation):
 
         return self.RProcess(self._devicemanager, cmd, stdout, stderr, env, cwd)
 
-    # be careful here as this inner class doesn't have access to outer class members    
+    # be careful here as this inner class doesn't have access to outer class members
     class RProcess(object):
         # device manager process
         dm = None
         def __init__(self, dm, cmd, stdout = None, stderr = None, env = None, cwd = None):
             self.dm = dm
             self.stdoutlen = 0
+            self.lastTestSeen = "remoteautomation.py"
             self.proc = dm.launchProcess(cmd, stdout, cwd, env, True)
             if (self.proc is None):
               if cmd[0] == 'am':
@@ -129,27 +152,73 @@ class RemoteAutomation(Automation):
             exepath = cmd[0]
             name = exepath.split('/')[-1]
             self.procName = name
+            # Hack for Robocop: Derive the actual process name from the command line.
+            # We expect something like:
+            #  ['am', 'instrument', '-w', '-e', 'class', 'org.mozilla.fennec.tests.testBookmark', 'org.mozilla.roboexample.test/android.test.InstrumentationTestRunner']
+            # and want to derive 'org.mozilla.fennec'.
+            if cmd[0] == 'am' and cmd[1] == "instrument":
+              try:
+                i = cmd.index("class")
+              except ValueError:
+                # no "class" argument -- maybe this isn't robocop?
+                i = -1
+              if (i > 0):
+                classname = cmd[i+1]
+                parts = classname.split('.')
+                try:
+                  i = parts.index("tests")
+                except ValueError:
+                  # no "tests" component -- maybe this isn't robocop?
+                  i = -1
+                if (i > 0):
+                  self.procName = '.'.join(parts[0:i])
+                  print "Robocop derived process name: "+self.procName
 
             # Setting timeout at 1 hour since on a remote device this takes much longer
             self.timeout = 3600
-            time.sleep(15)
+            # The benefit of the following sleep is unclear; it was formerly 15 seconds
+            time.sleep(1)
 
         @property
         def pid(self):
-            hexpid = self.dm.processExist(self.procName)
-            if (hexpid == None):
-                hexpid = "0x0"
-            return int(hexpid, 0)
-    
+            pid = self.dm.processExist(self.procName)
+            # HACK: we should probably be more sophisticated about monitoring
+            # running processes for the remote case, but for now we'll assume
+            # that this method can be called when nothing exists and it is not
+            # an error
+            if pid is None:
+                return 0
+            return pid
+
         @property
         def stdout(self):
-            t = self.dm.getFile(self.proc)
-            if t == None: return ''
-            tlen = len(t)
-            retVal = t[self.stdoutlen:]
-            self.stdoutlen = tlen
-            return retVal.strip('\n').strip()
- 
+            """ Fetch the full remote log file using devicemanager and return just
+                the new log entries since the last call (as a multi-line string).
+            """
+            if self.dm.fileExists(self.proc):
+                try:
+                    t = self.dm.pullFile(self.proc)
+                except DMError:
+                    # we currently don't retry properly in the pullFile
+                    # function in dmSUT, so an error here is not necessarily
+                    # the end of the world
+                    return ''
+                newLogContent = t[self.stdoutlen:]
+                self.stdoutlen = len(t)
+                # Match the test filepath from the last TEST-START line found in the new
+                # log content. These lines are in the form:
+                # 1234 INFO TEST-START | /filepath/we/wish/to/capture.html\n
+                testStartFilenames = re.findall(r"TEST-START \| ([^\s]*)", newLogContent)
+                if testStartFilenames:
+                    self.lastTestSeen = testStartFilenames[-1]
+                return newLogContent.strip('\n').strip()
+            else:
+                return ''
+
+        @property
+        def getLastTestSeen(self):
+            return self.lastTestSeen
+
         def wait(self, timeout = None):
             timer = 0
             interval = 5
@@ -165,9 +234,12 @@ class RemoteAutomation(Automation):
                 if (timer > timeout):
                     break
 
+            # Flush anything added to stdout during the sleep
+            print self.stdout
+
             if (timer >= timeout):
                 return 1
             return 0
- 
+
         def kill(self):
             self.dm.killProcess(self.procName)
